Add GNU printf attribute to kdb_dis_fprintf
GCC 13 flags this as a warning, which then becomes an error. This is probably
overkill for error checking but we're fixing it anyways. The attribute marks
the function as having printf-like capabilities and will enable additional
checks on usages of it to ensure formatting is correct to a printf-like
function. That didn't sound like a bad thing.