drm/xe/xe3p_xpc: Obtain dss/slice from fusing
On Xe3p, the physical number of DSS that compose a "slice" can vary by
platform SKU (note:  this is the group size _before_ we subtract any
fused-off DSS, not a number of DSS that are actually available/usable).
We should now read this number from the FUSE4 register and utilize it
when figuring out grpid and instanceid for steering purposes.

Bspec: 62483
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>