drm/xe/sim/fsg: Avoid problematic FS1 codepaths
As an experiment, add a fake workaround that will match on Intel's
current FCS simulation (which is the FSG platform rather than FS1).
When Habana simulation comes online, we'll likely add it as a different
class of presi, which will help distinguish between the two and won't
match this fake workaround.

If the "fake workaround" approach works for avoiding FS1 logic that will
crash the FSG simulators, then this may allow us to drop the separate
definitions of FS1 and FSG (and Xe3_XPC vs Xe3p_XPC) that we currently
have in the driver.

Signed-off-by: Matt Roper <matthew.d.roper@intel.com>