KVM: selftest: Fix the KVM_TDX_INIT_VM failed issue
Because the PASID XSAVE bit is supported after enqcmd patches are merged
and kselftest didn't filter it with the XFAM_FIXED0, it will set PASID bit in
the init_vm->cpuid. TDX module will report error as PASID should be fixed to 0.

Opportunistically modify the struct kvm_tdx_init_vm in the tools directory.

Signed-off-by: Chenyi Qiang <chenyi.qiang@intel.com>