KVM: TDX: Fix the check of running TDs when offlining CPUs
tdx_offline_cpu() is to prevent offlining the last CPU of a socket
if any TD is running (to reclaim HKIDs).

But the below check in tdx_offline_cpu():

	/* No TD is running.  Allow any cpu to be offline.  */
	if (atomic_read(&nr_configured_hkid) > 0)
		return 0;

iterally allows the last CPU offlining when TD is running.

commit a863d19f340a mistakenly modified the non-zero check to this way
when extending the nr_configured_hkid to indicate ongoing TDX module
updates.  so, just revert to the original check.

Note that when nr_configured_hkid() is negative, there maybe running TDs
e.g., during TD-preserving updates. In this case, offlining the last CPU
of a socket also should be rejected.

Alternative, we can use a separate flag instead of piggybacking on
nr_configured_hkid.

Fixes: a863d19f340a ("KVM: TDX: Add TDX module update support")
Signed-off-by: Chao Gao <chao.gao@intel.com>