x86/microcode: Allow fetching of ucode loaded from BIOS
This commit allows fetching a production microcode that
might not be commited by kernel but loaded from BIOS.
This is useful during `reload_nc` flow which requires
a committed ucode to be present in case the user wants
to do a rollback. This commit will automatically fetch the
production ucode from the standard production path
"/lib/firmware/" path if it doesn't find committed ucode
from the kernel.

This patch is optional can be removed before upstreaming.

Signed-off-by: Vijay Dhanraj <vijay.dhanraj@intel.com>