KVM: x86: Fix poll command
JIRA: https://issues.redhat.com/browse/RHEL-14702

According to the hardware manual, when the Poll command is issued, the
byte returned by the I/O read is 1 in Bit 7 when there is an interrupt,
and the highest priority binary code in Bits 2:0. The current pic
simulation code is not implemented strictly according to the above
expression.

Fix the implementation of pic_poll_read(), set Bit 7 when there is an
interrupt.

Signed-off-by: Jinliang Zheng <alexjlzheng@tencent.com>
Link: https://lore.kernel.org/r/20230419021924.1342184-1-alexjlzheng@tencent.com
Signed-off-by: Sean Christopherson <seanjc@google.com>
(cherry picked from commit 0d42522bdee70b9197be63dd76c9f6297cd1e832)
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>