KVM: SVM: Don't try to pointlessly single-step SEV-ES guests for NMI window
JIRA: https://issues.redhat.com/browse/RHEL-14702

Bail early from svm_enable_nmi_window() for SEV-ES guests without trying
to enable single-step of the guest, as single-stepping an SEV-ES guest is
impossible and the guest is responsible for *telling* KVM when it is ready
for an new NMI to be injected.

Functionally, setting TF and RF in svm->vmcb->save.rflags is benign as the
field is ignored by hardware, but it's all kinds of confusing.

Signed-off-by: Alexey Kardashevskiy <aik@amd.com>
Link: https://lore.kernel.org/r/20230615063757.3039121-10-aik@amd.com
Signed-off-by: Sean Christopherson <seanjc@google.com>
(cherry picked from commit a6bb5709029756a6151afed0eec26ba2bfabed51)
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>