dmaengine: idxd: Change wmb() to smp_wmb() when copying completion record to user space
wmb() is used to ensure status in the completion record is written
after the rest of the completion record is written and visible.
But since this memory barrier is not for SMP, user app polling the
status running on a different CPU may see the status before the reset
of the completion record.

Considering this scenario that event log handler is running on CPU1 while
user app polling completion record (cr) status is running on CPU2:

	CPU1				CPU2
event log handler			user app

					1. cr = 0 (status = 0)
2. copy X to user cr except "status"
3. wmb()
4. copy Y to user cr "status"
					5. poll status value Y
				 	6. read rest cr which is still 0.
					   cr handling fails
					7. cr value X visible now

Although wmb() ensure value Y is written and visible after X is written
on CPU1, the order is not guaranteed on CPU2. So user app may see status
value Y while cr value X is still not visible yet on CPU2. This will
cause reading 0 from the rest of cr and cr handling fails.

Changing wmb() to smp_wmb() ensures Y is written after X is written to cr
on both CPU1 and CPU2 and thus user app can consume cr in right order.

Signed-off-by: Fenghua Yu <fenghua.yu@intel.com>