KVM: X86: add a umsr kernel module
A mechanism is needed to present the user msr access to user space
application.

There can be multiple mechanism to achieve it.  A kernel module is the
easist and also naive way to achieve it. It can be used to verify the
Simics user msr support and KVM user msr support. A formal solution is
needed when we want to push the user msr change to upstream, which has
no plan yet.

Signed-off-by: Yunhong Jiang <yunhong.jiang@linux.intel.com>